backend/pyproject.toml
[project]
name = "knowledge-management-backend"
version = "1.0.0"
description = "A smart knowledge management system with AI-powered categorization"
requires-python = ">=3.10"

dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "networkx>=3.2.0",
    "openai>=1.3.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.6",
    "httpx>=0.25.0",
    "beautifulsoup4>=4.12.0",
    "markdown>=3.5.0",
    "pyyaml>=6.0.0",
    "chromadb>=0.4.0",
    "sentence-transformers>=2.2.0",
    "aiofiles>=24.1.0",
    "smolagents[toolkit,litellm]>=1.0.0",
    "watchdog>=6.0.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true 
